+++ compute_space/src/compute_space/web/app.py
@@ -18,6 +18,7 @@ from compute_space.core.startup import init_app
 from compute_space.core.terminal import cleanup_all as cleanup_terminal
 from compute_space.db import close_db
 from compute_space.db import get_db
+from compute_space.web.routes.pages.setup import setup_bp
 
 # ─── App Factory ───
 
@@ -42,10 +43,9 @@ def create_app(config: Config | None = None) -> Quart:
     app.teardown_appcontext(close_db)
 
     # Register blueprints (imported here per Flask/Quart convention - blueprints have globals/side effects)
-    from compute_space.web.auth.identity_routes import identity_bp  # noqa: PLC0415
-    from compute_space.web.auth.pages import auth_bp  # noqa: PLC0415
     from compute_space.web.routes.api.apps import api_apps_bp  # noqa: PLC0415
     from compute_space.web.routes.api.archive_backend import api_archive_backend_bp  # noqa: PLC0415
+    from compute_space.web.routes.api.identity import identity_bp  # noqa: PLC0415
     from compute_space.web.routes.api.permissions import api_permissions_bp  # noqa: PLC0415
     from compute_space.web.routes.api.permissions_v2 import api_permissions_v2_bp  # noqa: PLC0415
     from compute_space.web.routes.api.services import api_services_bp  # noqa: PLC0415
@@ -54,6 +54,7 @@ def create_app(config: Config | None = None) -> Quart:
     from compute_space.web.routes.api.system import system_bp  # noqa: PLC0415
     from compute_space.web.routes.docs import docs_bp  # noqa: PLC0415
     from compute_space.web.routes.pages.apps import apps_bp  # noqa: PLC0415
+    from compute_space.web.routes.pages.login import auth_bp  # noqa: PLC0415
     from compute_space.web.routes.pages.permissions import pages_permissions_bp  # noqa: PLC0415
     from compute_space.web.routes.pages.permissions_v2 import pages_permissions_v2_bp  # noqa: PLC0415
     from compute_space.web.routes.pages.settings import pages_settings_bp  # noqa: PLC0415
@@ -75,6 +76,7 @@ def create_app(config: Config | None = None) -> Quart:
     app.register_blueprint(api_permissions_bp)
     app.register_blueprint(services_bp)
     app.register_blueprint(services_v2_bp)
+    app.register_blueprint(setup_bp)
     app.register_blueprint(api_services_v2_bp)
     app.register_blueprint(api_permissions_v2_bp)
     app.register_blueprint(pages_permissions_v2_bp)
